+++ lib/arachni/plugin/manager.rb
@@ -60,18 +60,15 @@ class Manager < Arachni::Component::Manager
     # @raise [Error::UnsatisfiedDependency]
     #   If the environment is {#sane_env? not sane}.
     def run
-        prepared_signals = Queue.new
+        print_status 'Preparing plugins...'
 
         schedule.each do |name, options|
+            instance = create( name, options )
+            instance.prepare
+
             @jobs[name] = Thread.new do
                 exception_jail( false ) do
-                    begin
-                        Thread.current[:instance] = create( name, options )
-                        Thread.current[:instance].prepare
-                    ensure
-                        prepared_signals << nil
-                    end
-
+                    Thread.current[:instance] = instance
                     Thread.current[:instance].run
                     Thread.current[:instance].clean_up
 
@@ -82,10 +79,6 @@ class Manager < Arachni::Component::Manager
             end
         end
 
-        return if @jobs.empty?
-
-        print_status 'Preparing plugins...'
-        self.size.times { prepared_signals.pop }
         print_status '... done.'
     end
 
@@ -111,6 +104,7 @@ class Manager < Arachni::Component::Manager
 
         ordered << unordered
         ordered.flatten!
+        ordered.compact!
 
         ordered.inject({}) do |h, ph|
             name   = ph.keys.first